// src/types/inventory.ts
export interface InventoryItem {
  id: number
  magazine: string
  da: number
  price: string
  organic_traffic: string
  indexed_or_not: string
  do_follow_or_not: string
  category_section: string
  google_news: string
  sponsored: string
  cbd: string
  casino: string
  dating: string
  crypto: string
  contact: string
  created_at: string
  last_active_state: string
}